Move license check to own module
